refactor(framework) Move utils functions out of `serde.py` (#5343)
