+++ framework/py/flwr/common/constant.py
@@ -131,6 +131,9 @@ GC_THRESHOLD = 200_000_000  # 200 MB
 HEAD_BODY_DIVIDER = b"\x00"
 TYPE_BODY_LEN_DIVIDER = " "
 
+# Constants for serialization
+INT64_MAX_VALUE = 9223372036854775807  # (1 << 63) - 1
+
 
 class MessageType:
     """Message type."""
